% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gquad-package.R
\docType{package}
\name{gquad-package}
\alias{gquad-package}
\title{gquad-package: A package for predicting G quadruplex motif(s)}
\description{
The gquad-package provides two funcions for  prediction: gquad and gquadO.
gquad predicts without including overlaps, while gquadO includes overlaps.
Nucleotide sequence can be provided in raw or fasta format or as GenBank accession number(s).
Internet is needed to connect to GenBank database, if accession number(s) is given as argument.
A dataframe of G quadruplex motif(s) position, sequence and length is returned.
If more than one nucleotide sequence is provided as argument, an input ID is returned for motif(s) predicted from each input sequence.
}
\author{
Hannah O. Ajoge
}
\references{
paper under review
}

